refactor : 프로젝트 구조에 맞게 일부 수정
