+++ utils/codefDecoder.ts
@@ -3,34 +3,22 @@
  * URL 인코딩된 응답 데이터를 디코딩하고 JSON 파싱
  */
 
+import { GetRealEstateResponse } from '@/backend/realEstate/applications/dtos/RealEstateResponse';
+
 /**
- * 응답 데이터의 URL 디코딩 및 JSON 파싱 처리
- * @param data 원본 응답 데이터 (인코딩된 형태)
- * @returns 디코딩된 응답 데이터
+ * URL 인코딩된 문자열을 디코딩합니다.
+ * @param encodedString URL 인코딩된 문자열
+ * @returns 디코딩된 문자열
  */
-export function decodeCodefResponse(data: any) {
+export function decodeCodefResponse(
+  encodedString: string
+): GetRealEstateResponse {
   try {
-    // data 필드가 URL 인코딩된 문자열인 경우 디코딩
-    if (data) {
-      try {
-        // URL 디코딩
-        const urlDecodedData = decodeURIComponent(data);
-
-        // JSON 파싱
-        const jsonData = JSON.parse(urlDecodedData);
-
-        return jsonData;
-      } catch (decodeError) {
-        console.warn(
-          '⚠️ URL 디코딩 또는 JSON 파싱 실패, 원본 데이터 유지:',
-          decodeError
-        );
-      }
-    }
-
-    return data;
+    const urlDecodedData = decodeURIComponent(encodedString);
+    const jsonData = JSON.parse(urlDecodedData);
+    return jsonData as unknown as GetRealEstateResponse;
   } catch (error) {
-    console.error('❌ 응답 디코딩 처리 실패:', error);
-    return data; // 오류 시 원본 응답 반환
+    console.error('URL 디코딩 실패:', error);
+    return encodedString as unknown as GetRealEstateResponse; // 디코딩 실패 시 원본 반환
   }
 }
